All empty elements in fixvec are not allowed
